refactor(mitm-addon): split x connector usage tests
