--- crates/runner/mitm-addon/tests/test_x_connector_usage.py
@@ -1,1746 +1,0 @@
-"""Tests for direct X connector usage reporting."""
-
-import gzip
-import json
-import zlib
-from unittest.mock import patch
-
-import pytest
-
-import usage
-from body_limits import STREAM_BUFFER_LIMIT
-from tests.x_flow_helpers import (
-    json_body_that_exceeds_decoder_recursion,
-    json_body_that_exceeds_integer_digit_limit,
-    make_x_usage_flow,
-)
-
-
-class TestXConnectorUsage:
-    """Tests for report_connector_usage X billing behavior (issue #9504)."""
-
-    @pytest.fixture(autouse=True)
-    def _sync_executor(self, sync_usage_executor, usage_webhook_api):
-        """All tests here route billing through ``_call_and_get_billing`` which
-        inspects webhook delivery inline; the sync executor makes that work
-        without each test needing its own ``fresh_usage_executor`` + shutdown.
-        """
-        self._usage_webhook_api = usage_webhook_api
-
-    def _make_x_flow(
-        self,
-        real_flow,
-        tmp_path,
-        *,
-        path="/2/tweets",
-        query="",
-        body=b"",
-        status=200,
-        permission="tweet.read",
-        rule="GET /2/tweets",
-        content_encoding="",
-        request_body: bytes | None = None,
-        request_encoding: str | None = None,
-    ):
-        return make_x_usage_flow(
-            real_flow,
-            tmp_path,
-            path=path,
-            query=query,
-            body=body,
-            status=status,
-            permission=permission,
-            rule=rule,
-            content_encoding=content_encoding,
-            request_body=request_body,
-            request_encoding=request_encoding,
-        )
-
-    def _call_and_get_billing(self, flow, run_id="run-abc-123"):
-        """Call report_connector_usage and return the webhook payload(s).
-
-        Relies on the class-level ``_sync_executor`` autouse fixture to
-        route submissions inline.
-        """
-        with self._usage_webhook_api() as webhook:
-            start_count = webhook.request_count
-            usage.report_connector_usage(flow, run_id)
-            usage.flush_usage_events(trigger="test")
-        return [
-            event
-            for request in webhook.requests[start_count:]
-            for body in [request.json_body()]
-            for event in body["events"]
-        ]
-
-    def _call_and_get_single_billing(self, flow, run_id="run-abc-123"):
-        """Call report_connector_usage and return the single webhook payload."""
-        payloads = self._call_and_get_billing(flow, run_id)
-        assert len(payloads) == 1, f"expected 1 billing record, got {len(payloads)}"
-        return payloads[0]
-
-    def _assert_lost_visibility_error(self, proxy_log):
-        assert proxy_log.exists()
-        entries = [json.loads(line) for line in proxy_log.read_text().splitlines()]
-        matching_entries = [
-            entry
-            for entry in entries
-            if entry["level"] == "error" and "unparseable" in entry["message"].lower()
-        ]
-        assert len(matching_entries) == 1
-        return matching_entries[0]
-
-    def test_logs_single_resource_get(self, tmp_path, real_flow):
-        """GET /2/tweets/:id -> category=tweet.read, quantity=1."""
-        body = json.dumps({"data": {"id": "1", "text": "hi"}}).encode()
-        flow = self._make_x_flow(
-            real_flow, tmp_path, path="/2/tweets/1", body=body, rule="GET /2/tweets/{id}"
-        )
-        p = self._call_and_get_single_billing(flow)
-        assert p["category"] == "posts.read"
-        assert p["quantity"] == 1
-
-    def test_logs_batch_ids(self, tmp_path, real_flow):
-        """GET /2/tweets?ids=1,2,3 -> category=tweet.read, quantity=3."""
-        body = json.dumps({"data": [{"id": "1"}, {"id": "2"}, {"id": "3"}]}).encode()
-        flow = self._make_x_flow(real_flow, tmp_path, query="ids=1,2,3", body=body)
-        p = self._call_and_get_single_billing(flow)
-        assert p["category"] == "posts.read"
-        assert p["quantity"] == 3
-
-    def test_logs_batch_ids_with_deletions(self, tmp_path, real_flow):
-        """Batch with some missing ids -> bills actual data returned."""
-        body = json.dumps({"data": [{"id": "1"}, {"id": "3"}]}).encode()
-        flow = self._make_x_flow(real_flow, tmp_path, query="ids=1,2,3", body=body)
-        p = self._call_and_get_single_billing(flow)
-        assert p["category"] == "posts.read"
-        assert p["quantity"] == 2
-
-    def test_logs_expansions_includes(self, tmp_path, real_flow):
-        """?expansions=author_id -> three billing payloads for each resource type."""
-        body = json.dumps(
-            {
-                "data": [{"id": "1", "author_id": "99"}],
-                "includes": {
-                    "users": [{"id": "99"}],
-                    "media": [{"media_key": "m1"}, {"media_key": "m2"}],
-                },
-            }
-        ).encode()
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            query="expansions=author_id,attachments.media_keys",
-            body=body,
-        )
-        payloads = self._call_and_get_billing(flow)
-        by_cat = {p["category"]: p["quantity"] for p in payloads}
-        assert by_cat == {"posts.read": 1, "user.read": 1, "media.read": 2}
-
-    def test_posts_usage_events_as_one_batch(self, tmp_path, real_flow):
-        """One X response with multiple categories sends one batched webhook."""
-        body = json.dumps(
-            {
-                "data": [{"id": "1", "author_id": "99"}],
-                "includes": {"users": [{"id": "99"}]},
-            }
-        ).encode()
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            query="expansions=author_id",
-            body=body,
-        )
-
-        with self._usage_webhook_api() as webhook:
-            usage.report_connector_usage(flow, "run-abc-123")
-            assert webhook.request_count == 0
-            usage.flush_usage_events(trigger="test")
-
-        assert webhook.request_count == 1
-        [payload] = webhook.json_bodies()
-        assert payload["runId"] == "run-abc-123"
-        assert "idempotencyKey" not in payload
-        by_cat = {event["category"]: event for event in payload["events"]}
-        assert set(by_cat) == {"posts.read", "user.read"}
-        assert by_cat["posts.read"]["kind"] == "connector"
-        assert by_cat["posts.read"]["provider"] == "x"
-        assert by_cat["posts.read"]["quantity"] == 1
-        assert by_cat["user.read"]["quantity"] == 1
-
-    def test_source_dedupe_handles_colon_bearing_run_and_flow_ids(self, tmp_path, real_flow):
-        body = json.dumps({"data": {"id": "1", "text": "hi"}}).encode()
-        first = self._make_x_flow(
-            real_flow, tmp_path, path="/2/tweets/1", body=body, rule="GET /2/tweets/{id}"
-        )
-        second = self._make_x_flow(
-            real_flow, tmp_path, path="/2/tweets/2", body=body, rule="GET /2/tweets/{id}"
-        )
-        first.id = "flow"
-        second.id = "a:flow"
-
-        with self._usage_webhook_api() as webhook:
-            usage.report_connector_usage(first, "run:a")
-            usage.report_connector_usage(second, "run")
-            usage.flush_usage_events(trigger="test")
-
-        events_by_run = {body["runId"]: body["events"] for body in webhook.json_bodies()}
-        assert set(events_by_run) == {"run", "run:a"}
-        assert sum(event["quantity"] for events in events_by_run.values() for event in events) == 2
-        assert {
-            (event["kind"], event["provider"], event["category"], event["quantity"])
-            for events in events_by_run.values()
-            for event in events
-        } == {("connector", "x", "posts.read", 1)}
-
-    def test_bounds_unknown_include_categories_before_buffering(self, tmp_path, real_flow):
-        """Unknown includes cannot create unbounded synthetic usage categories."""
-        body = json.dumps(
-            {
-                "data": [],
-                "includes": {f"future_{index}": [{"id": str(index)}] for index in range(101)},
-            }
-        ).encode()
-        flow = self._make_x_flow(real_flow, tmp_path, query="expansions=future", body=body)
-
-        with self._usage_webhook_api() as webhook:
-            usage.report_connector_usage(flow, "run-abc-123")
-            usage.flush_usage_events(trigger="test")
-
-        bodies = webhook.json_bodies()
-        assert [len(body["events"]) for body in bodies] == [65]
-        assert {body["runId"] for body in bodies} == {"run-abc-123"}
-        assert all(
-            event["kind"] == "connector" and event["provider"] == "x"
-            for body in bodies
-            for event in body["events"]
-        )
-        by_cat = {
-            event["category"]: event["quantity"] for body in bodies for event in body["events"]
-        }
-        assert len(by_cat) == 65
-        assert "includes.future_0" in by_cat
-        assert "includes.future_63" in by_cat
-        assert "includes.future_64" not in by_cat
-        assert by_cat["includes.__overflow__"] == 37
-        assert all(len(category) <= 100 for category in by_cat)
-
-    def test_overlong_unknown_include_key_uses_overflow_category(self, tmp_path, real_flow):
-        """Unsafe synthetic categories are folded into the fallback-priced overflow bucket."""
-        overlong_key = "x" * 92
-        body = json.dumps(
-            {
-                "data": [{"id": "1"}],
-                "includes": {
-                    overlong_key: [{"id": "long"}],
-                    "bad/key": [{"id": "unsafe"}, {"id": "unsafe-2"}],
-                    "__overflow__": [{"id": "reserved"}],
-                },
-            }
-        ).encode()
-        flow = self._make_x_flow(real_flow, tmp_path, query="expansions=future", body=body)
-
-        payloads = self._call_and_get_billing(flow)
-        by_cat = {p["category"]: p["quantity"] for p in payloads}
-
-        assert by_cat == {"posts.read": 1, "includes.__overflow__": 4}
-        assert all(len(category) <= 100 for category in by_cat)
-
-    def test_empty_search_emits_no_billing(self, tmp_path, real_flow):
-        """Search returning zero results emits no usage_event row."""
-        body = json.dumps({"data": [], "meta": {"result_count": 0}}).encode()
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            path="/2/tweets/search/recent",
-            query="query=nothing",
-            body=body,
-            rule="GET /2/tweets/search/recent",
-        )
-        payloads = self._call_and_get_billing(flow)
-        assert payloads == []
-
-    def test_soft_error_with_request_hints_emits_no_billing(self, tmp_path, real_flow):
-        """HTTP 200 + errors array + no data field emits no usage_event
-        row (issue #9620)."""
-        body = json.dumps(
-            {
-                "errors": [
-                    {
-                        "value": "999999999999999999",
-                        "detail": "Could not find tweet with id: [999999999999999999].",
-                        "title": "Not Found Error",
-                        "resource_type": "tweet",
-                        "parameter": "id",
-                        "type": "https://api.twitter.com/2/problems/resource-not-found",
-                    }
-                ]
-            }
-        ).encode()
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            path="/2/tweets",
-            query="ids=999999999999999999",
-            body=body,
-            rule="GET /2/tweets",
-        )
-        payloads = self._call_and_get_billing(flow)
-        assert payloads == []
-
-    def test_zero_result_search_with_max_results_emits_no_billing(self, tmp_path, real_flow):
-        """Search with max_results=10 returning 0 results emits no
-        usage_event row (issue #9620)."""
-        body = json.dumps({"meta": {"result_count": 0, "newest_id": None}}).encode()
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            path="/2/tweets/search/recent",
-            query="query=xyzzy_no_results&max_results=10",
-            body=body,
-            rule="GET /2/tweets/search/recent",
-        )
-        payloads = self._call_and_get_billing(flow)
-        assert payloads == []
-
-    def test_logs_expansions_users_and_referenced_tweets(self, tmp_path, real_flow):
-        """includes.users and includes.tweets produce two billing payloads."""
-        body = json.dumps(
-            {
-                "data": [{"id": "1", "author_id": "99", "referenced_tweets": [{"id": "ref1"}]}],
-                "includes": {
-                    "users": [{"id": "99"}, {"id": "author2"}],
-                    "tweets": [{"id": "ref1"}, {"id": "ref2"}],
-                },
-            }
-        ).encode()
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            query="expansions=author_id,referenced_tweets.id",
-            body=body,
-        )
-        payloads = self._call_and_get_billing(flow)
-        by_cat = {p["category"]: p["quantity"] for p in payloads}
-        assert by_cat == {
-            "posts.read": 3,  # 1 primary + 2 referenced tweets
-            "user.read": 2,
-        }
-
-    def test_handles_unknown_includes_key(self, tmp_path, real_flow):
-        """Unknown includes.<key> types emit a synthetic ``includes.<key>``
-        category (the billing processor applies a server-side fallback
-        price) and emit a warn log at the decision point."""
-        sensitive_query = "vm0-sensitive-unknown-includes"
-        body = json.dumps(
-            {
-                "data": [{"id": "1"}],
-                "includes": {
-                    "users": [{"id": "99"}],
-                    "future_widget": [{"id": "w1"}, {"id": "w2"}, {"id": "w3"}],
-                },
-            }
-        ).encode()
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            path=f"/2/tweets?expansions=author_id&query={sensitive_query}",
-            body=body,
-        )
-        payloads = self._call_and_get_billing(flow)
-        by_cat = {p["category"]: p["quantity"] for p in payloads}
-        # 1 primary (posts.read) + 1 users include (mapped to user.read)
-        # + 3 unknown-widget includes (emitted as `includes.future_widget`).
-        assert by_cat == {
-            "posts.read": 1,
-            "user.read": 1,
-            "includes.future_widget": 3,
-        }
-        proxy_log = tmp_path / "proxy.jsonl"
-        assert proxy_log.exists()
-        content = proxy_log.read_text()
-        assert "future_widget" in content
-        assert "unrecognised" in content.lower()
-        assert '"level":"warn"' in content or '"level": "warn"' in content
-        assert sensitive_query not in content
-        entries = [json.loads(line) for line in content.splitlines()]
-        matching_entries = [
-            entry
-            for entry in entries
-            if entry.get("level") == "warn" and "unrecognised" in entry.get("message", "").lower()
-        ]
-        assert len(matching_entries) == 1
-        entry = matching_entries[0]
-        assert entry["url"] == "https://api.x.com/2/tweets"
-
-    def test_logs_search_meta_result_count(self, tmp_path, real_flow):
-        """Search response with meta.result_count -> quantity=20."""
-        body = json.dumps(
-            {
-                "data": [{"id": str(i)} for i in range(20)],
-                "meta": {"result_count": 20, "next_token": "abc"},
-            }
-        ).encode()
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            path="/2/tweets/search/recent",
-            query="query=hello&max_results=100",
-            body=body,
-            permission="tweet.read",
-            rule="GET /2/tweets/search/recent",
-        )
-        p = self._call_and_get_single_billing(flow)
-        assert p["category"] == "posts.read"
-        assert p["quantity"] == 20
-
-    def test_logs_users_by_usernames_batch(self, tmp_path, real_flow):
-        """GET /2/users/by?usernames=a,b,c -> category=users.read, quantity=2."""
-        body = json.dumps(
-            {"data": [{"id": "1", "username": "a"}, {"id": "2", "username": "b"}]}
-        ).encode()
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            path="/2/users/by",
-            query="usernames=a,b,c",
-            body=body,
-            permission="users.read",
-            rule="GET /2/users/by",
-        )
-        p = self._call_and_get_single_billing(flow)
-        assert p["category"] == "user.read"
-        assert p["quantity"] == 2
-
-    def test_logs_tweet_counts_total_tweet_count(self, tmp_path, real_flow):
-        """GET /2/tweets/counts/recent -> category=tweet.read, quantity=12567."""
-        body = json.dumps(
-            {
-                "data": [
-                    {"start": "2026-04-14T00:00", "end": "2026-04-15T00:00", "tweet_count": 8000},
-                    {"start": "2026-04-15T00:00", "end": "2026-04-16T00:00", "tweet_count": 4567},
-                ],
-                "meta": {"total_tweet_count": 12567},
-            }
-        ).encode()
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            path="/2/tweets/counts/recent",
-            query="query=hello",
-            body=body,
-            rule="GET /2/tweets/counts/recent",
-        )
-        p = self._call_and_get_single_billing(flow)
-        assert p["category"] == "posts.read"
-        assert p["quantity"] == 12567
-
-    def test_tweet_counts_zero_total_does_not_bill_buckets(self, tmp_path, real_flow):
-        """Count endpoint data arrays are time buckets, not returned posts."""
-        body = json.dumps(
-            {
-                "data": [
-                    {"start": "2026-04-14T00:00", "end": "2026-04-15T00:00", "tweet_count": 0},
-                    {"start": "2026-04-15T00:00", "end": "2026-04-16T00:00", "tweet_count": 0},
-                ],
-                "meta": {"total_tweet_count": 0},
-            }
-        ).encode()
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            path="/2/tweets/counts/recent",
-            query="query=nothing",
-            body=body,
-            rule="GET /2/tweets/counts/recent",
-        )
-
-        assert self._call_and_get_billing(flow) == []
-
-    @pytest.mark.parametrize(
-        "path",
-        ["/2/tweets/counts/recent", "/2/tweets/counts/all"],
-    )
-    def test_tweet_counts_total_lower_than_bucket_count_bills_total(
-        self, tmp_path, real_flow, path
-    ):
-        """Both count endpoints bill total_tweet_count, not time-bucket count."""
-        body = json.dumps(
-            {
-                "data": [
-                    {"start": "2026-04-14T00:00", "end": "2026-04-15T00:00", "tweet_count": 1},
-                    {"start": "2026-04-15T00:00", "end": "2026-04-16T00:00", "tweet_count": 0},
-                    {"start": "2026-04-16T00:00", "end": "2026-04-17T00:00", "tweet_count": 0},
-                ],
-                "meta": {"total_tweet_count": 1},
-            }
-        ).encode()
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            path=path,
-            query="query=rare",
-            body=body,
-            rule=f"GET {path}",
-        )
-
-        p = self._call_and_get_single_billing(flow)
-        assert p["category"] == "posts.read"
-        assert p["quantity"] == 1
-
-    def test_tweet_counts_path_matching_requires_exact_endpoint(self, tmp_path, real_flow):
-        body = json.dumps(
-            {
-                "data": [{"id": "1"}, {"id": "2"}, {"id": "3"}],
-                "meta": {"total_tweet_count": 1},
-            }
-        ).encode()
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            path="/2/tweets/counts/recent/extra",
-            body=body,
-            rule="GET /2/tweets/counts/recent/extra",
-        )
-
-        p = self._call_and_get_single_billing(flow)
-        assert p["category"] == "posts.read"
-        assert p["quantity"] == 3
-
-    def test_tweet_counts_missing_total_skips_billing_and_logs_warning(self, tmp_path, real_flow):
-        """Parsed count endpoint responses without total_tweet_count should not bill buckets."""
-        body = json.dumps(
-            {
-                "data": [
-                    {"start": "2026-04-14T00:00", "end": "2026-04-15T00:00", "tweet_count": 2},
-                    {"start": "2026-04-15T00:00", "end": "2026-04-16T00:00", "tweet_count": 3},
-                ],
-                "meta": {},
-            }
-        ).encode()
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            path="/2/tweets/counts/recent",
-            query="query=missing_total",
-            body=body,
-            rule="GET /2/tweets/counts/recent",
-        )
-
-        assert self._call_and_get_billing(flow) == []
-
-        proxy_log = tmp_path / "proxy.jsonl"
-        entries = [json.loads(line) for line in proxy_log.read_text().splitlines()]
-        assert any(
-            entry["level"] == "warn" and "total_tweet_count" in entry["message"]
-            for entry in entries
-        )
-
-    def test_tweet_counts_unparseable_ignores_request_hints_and_logs_error(
-        self, tmp_path, real_flow
-    ):
-        """Count endpoint request hints do not represent returned post count."""
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            path="/2/tweets/counts/recent",
-            query="max_results=50",
-            body=b"not json",
-            rule="GET /2/tweets/counts/recent",
-        )
-
-        assert self._call_and_get_billing(flow) == []
-
-        proxy_log = tmp_path / "proxy.jsonl"
-        entries = [json.loads(line) for line in proxy_log.read_text().splitlines()]
-        assert any(
-            entry["level"] == "error" and "count endpoint" in entry["message"] for entry in entries
-        )
-
-    def test_handles_gzip_body(self, tmp_path, real_flow):
-        """gzip-encoded response body decompresses before parsing."""
-        raw = json.dumps({"data": [{"id": "1"}], "meta": {"result_count": 1}}).encode()
-        body = gzip.compress(raw)
-        flow = self._make_x_flow(real_flow, tmp_path, body=body, content_encoding="gzip")
-        p = self._call_and_get_single_billing(flow)
-        assert p["category"] == "posts.read"
-        assert p["quantity"] == 1
-
-    def test_logs_write_operation_charges_one(self, tmp_path, real_flow):
-        """POST /2/tweets (no request body parsed) -> stay on the expensive
-        with_url bucket, quantity=1."""
-        body = json.dumps({"data": {"id": "99", "text": "new tweet"}}).encode()
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            path="/2/tweets",
-            body=body,
-            status=201,
-            permission="tweet.write",
-            rule="POST /2/tweets",
-        )
-        flow.request.method = "POST"
-        p = self._call_and_get_single_billing(flow)
-        assert p["category"] == "content.create_with_url"
-        assert p["quantity"] == 1
-
-    def test_x_json_parse_error_on_write_does_not_emit_lost_visibility_log(
-        self, tmp_path, real_flow
-    ):
-        """Write operations bill by method and should not emit read visibility errors."""
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            path="/2/tweets",
-            status=201,
-            permission="tweet.write",
-            rule="POST /2/tweets",
-        )
-        flow.request.method = "POST"
-        flow.metadata["x_json_state"] = {
-            "body_parsed": False,
-            "body_truncated": False,
-            "parse_error": "incomplete json",
-        }
-        proxy_log = tmp_path / "proxy.jsonl"
-
-        p = self._call_and_get_single_billing(flow)
-
-        assert p["category"] == "content.create_with_url"
-        assert p["quantity"] == 1
-        assert proxy_log.exists()
-        entries = [json.loads(line) for line in proxy_log.read_text().splitlines()]
-        assert all(entry["level"] != "error" for entry in entries)
-        assert all("unparseable" not in entry["message"].lower() for entry in entries)
-        assert all("parse_error" not in entry for entry in entries)
-
-    def test_tweet_create_plain_text_downgrades_to_content_create(self, tmp_path, real_flow):
-        """POST /2/tweets with text only (no URL, no quote, no media)
-        downgrades to the cheaper Content: Create bucket."""
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            path="/2/tweets",
-            body=json.dumps({"data": {"id": "1"}}).encode(),
-            status=201,
-            permission="tweet.write",
-            rule="POST /2/tweets",
-        )
-        flow.request.method = "POST"
-        flow.request.content = json.dumps({"text": "hello world"}).encode()
-        p = self._call_and_get_single_billing(flow)
-        assert p["category"] == "content.create"
-        assert p["quantity"] == 1
-
-    @pytest.mark.parametrize(
-        ("request_encoding", "request_body"),
-        [
-            ("gzip", gzip.compress(json.dumps({"text": "hello world"}).encode())),
-            ("deflate", zlib.compress(json.dumps({"text": "hello world"}).encode())),
-        ],
-    )
-    def test_tweet_create_compressed_plain_text_downgrades_to_content_create(
-        self, tmp_path, real_flow, request_encoding, request_body
-    ):
-        """Small compressed tweet create bodies still refine to Content: Create."""
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            path="/2/tweets",
-            body=json.dumps({"data": {"id": "1"}}).encode(),
-            status=201,
-            permission="tweet.write",
-            rule="POST /2/tweets",
-            request_body=request_body,
-            request_encoding=request_encoding,
-        )
-        flow.request.method = "POST"
-        p = self._call_and_get_single_billing(flow)
-        assert p["category"] == "content.create"
-        assert p["quantity"] == 1
-
-    def test_tweet_create_gzip_decoded_body_over_cap_stays_conservative(self, tmp_path, real_flow):
-        """A gzip body that expands beyond the billing inspection cap is not refined."""
-        long_text = "x" * STREAM_BUFFER_LIMIT
-        request_body = gzip.compress(json.dumps({"text": long_text}).encode())
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            path="/2/tweets",
-            body=json.dumps({"data": {"id": "1"}}).encode(),
-            status=201,
-            permission="tweet.write",
-            rule="POST /2/tweets",
-            request_body=request_body,
-            request_encoding="gzip",
-        )
-        flow.request.method = "POST"
-        p = self._call_and_get_single_billing(flow)
-        assert p["category"] == "content.create_with_url"
-        assert p["quantity"] == 1
-
-    def test_tweet_create_raw_body_over_cap_stays_conservative(self, tmp_path, real_flow):
-        """An oversized identity request body is not refined."""
-        request_body = b"{" + b'"text":"' + b"x" * STREAM_BUFFER_LIMIT + b'"}'
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            path="/2/tweets",
-            body=json.dumps({"data": {"id": "1"}}).encode(),
-            status=201,
-            permission="tweet.write",
-            rule="POST /2/tweets",
-            request_body=request_body,
-        )
-        flow.request.method = "POST"
-        p = self._call_and_get_single_billing(flow)
-        assert p["category"] == "content.create_with_url"
-        assert p["quantity"] == 1
-
-    @pytest.mark.parametrize("request_encoding", ["gzip", "br", "zstd", "x-vm0-test"])
-    def test_tweet_create_invalid_or_unsupported_encoding_stays_conservative(
-        self, tmp_path, real_flow, request_encoding
-    ):
-        """Invalid compressed or unsupported encoded bodies are not refined."""
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            path="/2/tweets",
-            body=json.dumps({"data": {"id": "1"}}).encode(),
-            status=201,
-            permission="tweet.write",
-            rule="POST /2/tweets",
-            request_body=b'{"text":"hello world"}',
-            request_encoding=request_encoding,
-        )
-        flow.request.method = "POST"
-        p = self._call_and_get_single_billing(flow)
-        assert p["category"] == "content.create_with_url"
-        assert p["quantity"] == 1
-
-    def test_non_refinement_flow_does_not_decode_request_body(self, tmp_path, real_flow):
-        """Only body-refinement candidates should inspect request content."""
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            path="/2/tweets/123/retweeted_by?max_results=10",
-            body=json.dumps({"data": [{"id": "u1"}]}).encode(),
-            permission="tweet.read",
-            rule="GET /2/tweets/{id}/retweeted_by",
-            request_body=b"not gzip request content",
-            request_encoding="gzip",
-        )
-        flow.metadata["original_url"] = "https://api.x.com/2/tweets/123/retweeted_by?max_results=10"
-
-        with patch(
-            "usage.providers.connectors.x.billing_body.decode_request_body_for_billing"
-        ) as decode_request_body:
-            p = self._call_and_get_single_billing(flow)
-
-        decode_request_body.assert_not_called()
-        assert p["category"] == "user.read"
-        assert p["quantity"] == 1
-
-    @pytest.mark.parametrize(
-        "text",
-        [
-            "check https://example.com",
-            "check HTTPS://example.com",
-            "check HtTp://www.ExaMPLE.COM/index.html",
-            "Visit vm0.ai for details",
-            "Visit vm0.ai.",
-            "(vm0.ai)",
-            "Visit go.dev",
-            "Read example.museum",
-            "Label:example.com",
-            "Param url=example.com",
-            "Pipe|example.com",
-            "Visit EXAMPLE.COM",
-            "Visit 123.com",
-            "Visit blog.example.co.uk",
-            "Open example.com/path/to/resource?search=foo&lang=en",
-            "Open example.com:443/path",
-            "Open xn--r8jz45g.xn--q9jyb4c",
-            "IDN 例え.みんな",
-            "Accent mañana.com",
-            "Sharp S faß.de",
-            "Fullwidth compatibility \uff26\uff2f\uff2f.com",
-        ],
-    )
-    def test_tweet_create_with_url_stays_on_with_url_bucket(self, tmp_path, real_flow, text):
-        """POST /2/tweets whose text contains a URL stays on the
-        Content: Create (with URL) bucket."""
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            path="/2/tweets",
-            body=json.dumps({"data": {"id": "1"}}).encode(),
-            status=201,
-            permission="tweet.write",
-            rule="POST /2/tweets",
-        )
-        flow.request.method = "POST"
-        flow.request.content = json.dumps({"text": text}).encode()
-        p = self._call_and_get_single_billing(flow)
-        assert p["category"] == "content.create_with_url"
-        assert p["quantity"] == 1
-
-    @pytest.mark.parametrize(
-        "text",
-        [
-            "Email support@example.com",
-            "Mention @twitter.com",
-            "Tag #twitter.com",
-            "Cash $twitter.com",
-            "Path /twitter.com",
-            "Archive long.test.tar.bz2",
-            "Word abcHTTPS://example.com",
-            "Fullwidth mention \uff20twitter.com",
-            "Fullwidth tag \uff03twitter.com",
-            "Plus suffix example.com+tag",
-            "At suffix example.com@user",
-            "Unknown example.notatld",
-            "Fullwidth unknown \uff26\uff2f\uff2f.notatld",
-            "Underscore foo_bar.example.com",
-            "Leading hyphen -bad.com",
-            "Trailing hyphen bad-.com",
-        ],
-    )
-    def test_tweet_create_url_like_non_links_downgrade_to_content_create(
-        self, tmp_path, real_flow, text
-    ):
-        """twitter-text-style boundary guards avoid obvious non-link
-        domains while still allowing plain text to downgrade."""
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            path="/2/tweets",
-            body=json.dumps({"data": {"id": "1"}}).encode(),
-            status=201,
-            permission="tweet.write",
-            rule="POST /2/tweets",
-        )
-        flow.request.method = "POST"
-        flow.request.content = json.dumps({"text": text}).encode()
-        p = self._call_and_get_single_billing(flow)
-        assert p["category"] == "content.create"
-        assert p["quantity"] == 1
-
-    def test_tweet_create_rendered_link_signals_stay_on_with_url_bucket(self, tmp_path, real_flow):
-        """Quote tweets, attached media, cards, and DM deep links render links;
-        we stay on the expensive bucket even when the text has no URL."""
-        for req_body in [
-            # quote tweet
-            json.dumps({"text": "nice", "quote_tweet_id": "abc"}).encode(),
-            # attached media
-            json.dumps({"text": "pic", "media": {"media_ids": ["42"]}}).encode(),
-            # attached card
-            json.dumps({"text": "card", "card_uri": "card://123"}).encode(),
-            # direct-message deep link
-            json.dumps(
-                {
-                    "text": "DM me for details",
-                    "direct_message_deep_link": "https://x.com/messages/compose?recipient_id=123",
-                }
-            ).encode(),
-        ]:
-            flow = self._make_x_flow(
-                real_flow,
-                tmp_path,
-                path="/2/tweets",
-                body=json.dumps({"data": {"id": "1"}}).encode(),
-                status=201,
-                permission="tweet.write",
-                rule="POST /2/tweets",
-            )
-            flow.request.method = "POST"
-            flow.request.content = req_body
-            p = self._call_and_get_single_billing(flow)
-            assert p["category"] == "content.create_with_url"
-            assert p["quantity"] == 1
-
-    @pytest.mark.parametrize("direct_message_deep_link", ["", "   ", None, 123, {"url": "x"}])
-    def test_tweet_create_invalid_direct_message_deep_link_downgrades_to_content_create(
-        self, tmp_path, real_flow, direct_message_deep_link
-    ):
-        """Invalid DM deep-link values do not block the plain-text downgrade."""
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            path="/2/tweets",
-            body=json.dumps({"data": {"id": "1"}}).encode(),
-            status=201,
-            permission="tweet.write",
-            rule="POST /2/tweets",
-        )
-        flow.request.method = "POST"
-        flow.request.content = json.dumps(
-            {"text": "DM me for details", "direct_message_deep_link": direct_message_deep_link}
-        ).encode()
-        p = self._call_and_get_single_billing(flow)
-        assert p["category"] == "content.create"
-        assert p["quantity"] == 1
-
-    def test_tweet_create_unparseable_body_stays_conservative(self, tmp_path, real_flow):
-        """A malformed request body keeps billing on the max bucket so
-        we never under-charge on parse failure."""
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            path="/2/tweets",
-            body=json.dumps({"data": {"id": "1"}}).encode(),
-            status=201,
-            permission="tweet.write",
-            rule="POST /2/tweets",
-        )
-        flow.request.method = "POST"
-        flow.request.content = b"not valid json at all"
-        p = self._call_and_get_single_billing(flow)
-        assert p["category"] == "content.create_with_url"
-
-    @pytest.mark.parametrize(
-        "request_body",
-        [
-            pytest.param(json_body_that_exceeds_decoder_recursion(), id="decoder-recursion"),
-            pytest.param(json_body_that_exceeds_integer_digit_limit(), id="integer-digit-limit"),
-        ],
-    )
-    def test_tweet_create_json_parser_failure_stays_conservative(
-        self, tmp_path, real_flow, request_body
-    ):
-        """Stdlib JSON parser failures must not interrupt tweet-create billing."""
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            path="/2/tweets",
-            body=json.dumps({"data": {"id": "1"}}).encode(),
-            status=201,
-            permission="tweet.write",
-            rule="POST /2/tweets",
-        )
-        flow.request.method = "POST"
-        flow.request.content = request_body
-
-        p = self._call_and_get_single_billing(flow)
-
-        assert p["category"] == "content.create_with_url"
-        assert p["quantity"] == 1
-
-    def test_query_string_does_not_break_literal_suffix_override(self, tmp_path, real_flow):
-        """``flow.request.path`` from mitmproxy includes the query string;
-        literal-suffix overrides (e.g. ``/2/tweets/{id}/retweeted_by``)
-        must still fire.  Regression guard for under-charging on
-        popular paginated read endpoints."""
-        body = json.dumps({"data": [{"id": "u1"}, {"id": "u2"}]}).encode()
-        # The helper sets flow.request.path = path (no query), so we
-        # craft the path-with-query explicitly to mirror what mitmproxy
-        # delivers in real traffic.
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            path="/2/tweets/123/retweeted_by?max_results=10",
-            body=body,
-            permission="tweet.read",
-            rule="GET /2/tweets/{id}/retweeted_by",
-        )
-        # The ?max_results metadata goes into original_url for
-        # req-meta parsing to consume.
-        flow.metadata["original_url"] = "https://api.x.com/2/tweets/123/retweeted_by?max_results=10"
-        p = self._call_and_get_single_billing(flow)
-        assert p["category"] == "user.read"
-        assert p["quantity"] == 2
-
-    def test_delete_method_charges_one(self, tmp_path, real_flow):
-        """DELETE /2/tweets/{id} routes to Content: Manage, not the
-        tweet.write scope default.  Writes always charge quantity=1
-        regardless of response shape."""
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            path="/2/tweets/123",
-            body=b"",
-            permission="tweet.write",
-            rule="DELETE /2/tweets/{id}",
-        )
-        flow.request.method = "DELETE"
-        p = self._call_and_get_single_billing(flow)
-        assert p["category"] == "content.manage"
-        assert p["quantity"] == 1
-
-    def test_expansion_with_empty_includes_array(self, tmp_path, real_flow):
-        """includes.users is empty array -> no users.read billing record."""
-        body = json.dumps(
-            {
-                "data": [{"id": "1"}],
-                "includes": {"users": []},
-            }
-        ).encode()
-        flow = self._make_x_flow(real_flow, tmp_path, query="expansions=author_id", body=body)
-        p = self._call_and_get_single_billing(flow)
-        assert p["category"] == "posts.read"
-        assert p["quantity"] == 1
-
-    def test_empty_search_with_includes_emits_only_includes(self, tmp_path, real_flow):
-        """Search returns 0 data but non-empty includes -> only the
-        includes row is emitted; the zero-primary row is skipped."""
-        body = json.dumps(
-            {
-                "data": [],
-                "meta": {"result_count": 0},
-                "includes": {"users": [{"id": "u1"}]},
-            }
-        ).encode()
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            path="/2/tweets/search/recent",
-            query="query=test&expansions=author_id",
-            body=body,
-            rule="GET /2/tweets/search/recent",
-        )
-        p = self._call_and_get_single_billing(flow)
-        assert p["category"] == "user.read"
-        assert p["quantity"] == 1
-
-    def test_logs_x_stream_with_ndjson_state(self, tmp_path, real_flow):
-        """Stream with pre-populated x_ndjson_state -> two billing payloads."""
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            path="/2/tweets/search/stream",
-            body=b"",
-            rule="GET /2/tweets/search/stream",
-        )
-        flow.metadata["x_ndjson_state"] = {
-            "data_count": 50,
-            "includes": {"users": 47, "tweets": 12},
-            "lines_parsed": 50,
-            "lines_failed": 1,
-        }
-        payloads = self._call_and_get_billing(flow)
-        by_cat = {p["category"]: p["quantity"] for p in payloads}
-        # tweet.read primary 50 + 12 from includes.tweets = 62
-        assert by_cat["posts.read"] == 62
-        assert by_cat["user.read"] == 47
-
-    def test_x_stream_empty_emits_no_billing(self, tmp_path, real_flow):
-        """Stream that delivered 0 tweets emits no usage_event row, and
-        in particular does NOT trigger _X_UNPARSEABLE_READ_FALLBACK."""
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            path="/2/tweets/search/stream",
-            body=b"",
-            rule="GET /2/tweets/search/stream",
-        )
-        flow.metadata["x_ndjson_state"] = {
-            "data_count": 0,
-            "includes": {},
-            "lines_parsed": 0,
-            "lines_failed": 0,
-        }
-        payloads = self._call_and_get_billing(flow)
-        assert payloads == []
-
-    def test_legacy_x_json_fallback_extracts_selective_field_counts(self, tmp_path, real_flow):
-        """Buffered fallback should share X JSON field semantics with the selective parser."""
-        body = json.dumps(
-            {
-                "data": [{"id": "1"}, {"id": "2"}],
-                "errors": [{"title": "partial failure"}],
-                "includes": {
-                    "users": [{"id": "u1"}],
-                    "media": [],
-                    "topics": "ignored",
-                },
-                "meta": {"result_count": 2, "total_tweet_count": 3},
-            }
-        ).encode()
-        flow = self._make_x_flow(real_flow, tmp_path, body=body)
-
-        payloads = self._call_and_get_billing(flow)
-
-        by_cat = {p["category"]: p["quantity"] for p in payloads}
-        assert by_cat == {"posts.read": 2, "user.read": 1}
-
-    def test_legacy_x_json_fallback_ignores_boolean_result_count(self, tmp_path, real_flow):
-        body = json.dumps({"meta": {"result_count": True}}).encode()
-        flow = self._make_x_flow(real_flow, tmp_path, body=body)
-        proxy_log = tmp_path / "proxy.jsonl"
-
-        payloads = self._call_and_get_billing(flow)
-
-        assert payloads == []
-        if proxy_log.exists():
-            entries = [json.loads(line) for line in proxy_log.read_text().splitlines()]
-            assert all(entry["level"] != "error" for entry in entries)
-            assert all("unparseable" not in entry["message"].lower() for entry in entries)
-
-    def test_truncated_buffer_with_no_hints_skips_billing(self, tmp_path, real_flow):
-        """Unparseable body + no URL hints: skip emission and log an
-        error.  The previous blind fallback of 100 units was removed;
-        ops audits via the proxy error log instead."""
-        flow = self._make_x_flow(real_flow, tmp_path, body=b"{")
-        flow.metadata["stream_buffer_state"] = {"truncated": True}
-        proxy_log = tmp_path / "proxy.jsonl"
-
-        assert self._call_and_get_billing(flow) == []
-
-        entry = json.loads(proxy_log.read_text().splitlines()[0])
-        assert entry["level"] == "error"
-        assert "unparseable" in entry["message"].lower()
-        assert entry["body_truncated"] is True
-        assert "parse_error" not in entry
-
-    def test_invalid_json_with_no_hints_skips_billing(self, tmp_path, real_flow):
-        """Malformed body + no URL hints: skip emission (see above)."""
-        flow = self._make_x_flow(real_flow, tmp_path, body=b"not json")
-        assert self._call_and_get_billing(flow) == []
-
-    def test_non_dict_json_with_no_hints_skips_billing(self, tmp_path, real_flow):
-        """A valid non-object JSON response preserves the old unparseable fallback."""
-        flow = self._make_x_flow(real_flow, tmp_path, body=b"[1,2,3]")
-        assert self._call_and_get_billing(flow) == []
-
-    def test_array_element_fields_do_not_drive_x_billing(self, tmp_path, real_flow):
-        """Fields under array elements must not masquerade as top-level X metadata."""
-        body = json.dumps(
-            {
-                "data": [],
-                "includes": [{"users": [{"id": "u1"}]}],
-                "meta": [{"result_count": 5}],
-            }
-        ).encode()
-        flow = self._make_x_flow(real_flow, tmp_path, body=body)
-        assert self._call_and_get_billing(flow) == []
-
-    def test_unparseable_no_hints_writes_error_to_proxy_log(self, tmp_path, real_flow):
-        """Operators must be able to audit the lost-visibility case:
-        the proxy log receives a structured error entry."""
-        sensitive_query = "vm0-sensitive-unparseable"
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            path=f"/2/tweets/search/recent?query={sensitive_query}",
-            body=b"not json",
-            rule="GET /2/tweets/search/recent",
-        )
-        flow.metadata["original_url"] = (
-            f"https://api.x.com:8443/2/tweets/search/recent?query={sensitive_query}"
-        )
-        proxy_log = tmp_path / "proxy.jsonl"
-        assert self._call_and_get_billing(flow) == []
-        assert proxy_log.exists()
-        content = proxy_log.read_text()
-        assert sensitive_query not in content
-        entries = [json.loads(line) for line in content.splitlines()]
-        matching_entries = [
-            entry
-            for entry in entries
-            if entry.get("level") == "error" and "unparseable" in entry.get("message", "").lower()
-        ]
-        assert len(matching_entries) == 1
-        entry = matching_entries[0]
-        assert entry["level"] == "error"
-        assert "unparseable" in entry["message"].lower()
-        assert entry["permission"] == "tweet.read"
-        assert entry["url"] == "https://api.x.com:8443/2/tweets/search/recent"
-        assert "parse_error" not in entry
-
-    def test_unparseable_x_json_state_logs_parse_error(self, tmp_path, real_flow):
-        """Incremental parser failures should surface the parse reason for audit."""
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            path="/2/tweets/search/recent",
-            rule="GET /2/tweets/search/recent",
-        )
-        flow.metadata["x_json_state"] = {
-            "body_parsed": False,
-            "body_truncated": False,
-            "parse_error": "incomplete json",
-        }
-        proxy_log = tmp_path / "proxy.jsonl"
-
-        assert self._call_and_get_billing(flow) == []
-
-        entry = json.loads(proxy_log.read_text().splitlines()[0])
-        assert entry["level"] == "error"
-        assert "unparseable" in entry["message"].lower()
-        assert entry["parse_error"] == "incomplete json"
-
-    @pytest.mark.parametrize(
-        "parse_error",
-        ["", "   ", None, b"incomplete json", {"reason": "incomplete json"}],
-    )
-    def test_unparseable_x_json_state_omits_invalid_parse_error(
-        self, tmp_path, real_flow, parse_error
-    ):
-        """Only non-empty string parse errors should be written to the audit log."""
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            path="/2/tweets/search/recent",
-            rule="GET /2/tweets/search/recent",
-        )
-        flow.metadata["x_json_state"] = {
-            "body_parsed": False,
-            "body_truncated": False,
-            "parse_error": parse_error,
-        }
-        proxy_log = tmp_path / "proxy.jsonl"
-
-        assert self._call_and_get_billing(flow) == []
-
-        entry = json.loads(proxy_log.read_text().splitlines()[0])
-        assert entry["level"] == "error"
-        assert "unparseable" in entry["message"].lower()
-        assert "parse_error" not in entry
-
-    def test_billable_counts_fallback_only_when_no_hints(self, tmp_path, real_flow):
-        """body unparseable but ?ids= present -> uses ids_count, no fallback."""
-        flow = self._make_x_flow(real_flow, tmp_path, query="ids=1,2,3", body=b"not json")
-        p = self._call_and_get_single_billing(flow)
-        assert p["category"] == "posts.read"
-        assert p["quantity"] == 3
-
-    def test_parsed_response_ignores_oversized_fallback_query(self, tmp_path, real_flow):
-        """Parsed responses bill from the body and do not need query fallback hints."""
-        body = json.dumps({"data": [{"id": "1"}, {"id": "2"}]}).encode()
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            query=f"ignored={'x' * 200_000}",
-            body=body,
-        )
-
-        p = self._call_and_get_single_billing(flow)
-
-        assert p["category"] == "posts.read"
-        assert p["quantity"] == 2
-
-    def test_unparseable_response_scans_relevant_fallback_query_hints(self, tmp_path, real_flow):
-        """Fallback query scanning ignores unrelated params without materializing them."""
-        irrelevant_params = "&".join(f"noise{i}=value{i}" for i in range(500))
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            query=f"{irrelevant_params}&ids=1,2,3",
-            body=b"not json",
-        )
-
-        p = self._call_and_get_single_billing(flow)
-
-        assert p["category"] == "posts.read"
-        assert p["quantity"] == 3
-
-    def test_unparseable_response_accumulates_repeated_id_like_fallback_hints(
-        self, tmp_path, real_flow
-    ):
-        """Repeated allowed selector keys are accumulated on the matching path."""
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            query="ids=1,2&ids=3",
-            body=b"not json",
-        )
-
-        p = self._call_and_get_single_billing(flow)
-
-        assert p["category"] == "posts.read"
-        assert p["quantity"] == 3
-
-    def test_unparseable_response_does_not_treat_semicolon_as_query_separator(
-        self, tmp_path, real_flow
-    ):
-        """Semicolons stay inside values, matching Python 3.12 parse_qs behavior."""
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            query="ids=1;max_results=50",
-            body=b"not json",
-        )
-
-        p = self._call_and_get_single_billing(flow)
-
-        assert p["category"] == "posts.read"
-        assert p["quantity"] == 1
-
-    def test_oversized_fallback_query_suppresses_request_hints(self, tmp_path, real_flow):
-        """Oversized fallback queries are treated as having no reliable count hints."""
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            query=f"ids=1,2,3&ignored={'x' * 200_000}",
-            body=b"not json",
-        )
-        proxy_log = tmp_path / "proxy.jsonl"
-
-        assert self._call_and_get_billing(flow) == []
-
-        assert proxy_log.exists()
-        content = proxy_log.read_text()
-        assert "unparseable" in content.lower()
-        assert '"level":"error"' in content or '"level": "error"' in content
-
-    def test_unencodable_fallback_query_suppresses_request_hints(self, tmp_path, real_flow):
-        """Malformed Unicode in fallback queries should fail closed instead of crashing."""
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            query="noise=" + "\ud800" + "&ids=1",
-            body=b"not json",
-        )
-        proxy_log = tmp_path / "proxy.jsonl"
-
-        assert self._call_and_get_billing(flow) == []
-        self._assert_lost_visibility_error(proxy_log)
-
-    @pytest.mark.parametrize(
-        ("path", "query_template", "permission", "rule", "expected_quantity"),
-        [
-            (
-                "/2/tweets/search/recent",
-                "ids={large}&query=hello&max_results=50",
-                "tweet.read",
-                "GET /2/tweets/search/recent",
-                50,
-            ),
-            (
-                "/2/tweets",
-                "max_results={large}&ids=1,2",
-                "tweet.read",
-                "GET /2/tweets",
-                2,
-            ),
-            (
-                "/2/tweets/search/recent",
-                "query=hello&max_results=50&max_results={large}",
-                "tweet.read",
-                "GET /2/tweets/search/recent",
-                50,
-            ),
-        ],
-    )
-    def test_unparseable_response_ignores_oversized_irrelevant_fallback_hint_values(
-        self,
-        tmp_path,
-        real_flow,
-        path,
-        query_template,
-        permission,
-        rule,
-        expected_quantity,
-    ):
-        """Oversized values only fail the hint that would actually be trusted."""
-        query = query_template.format(large="x" * 20_000)
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            path=path,
-            query=query,
-            body=b"not json",
-            permission=permission,
-            rule=rule,
-        )
-
-        p = self._call_and_get_single_billing(flow)
-
-        assert p["quantity"] == expected_quantity
-
-    @pytest.mark.parametrize(
-        ("path", "query", "permission", "rule", "expected_category", "expected_quantity"),
-        [
-            ("/2/tweets", "ids=1,2,3", "tweet.read", "GET /2/tweets", "posts.read", 3),
-            (
-                "/2/tweets/search/recent",
-                "query=hello&max_results=50",
-                "tweet.read",
-                "GET /2/tweets/search/recent",
-                "posts.read",
-                50,
-            ),
-            (
-                "/2/users/search",
-                "query=alice&max_results=1000",
-                "users.read",
-                "GET /2/users/search",
-                "user.read",
-                1000,
-            ),
-            ("/2/users", "ids=1,2", "users.read", "GET /2/users", "user.read", 2),
-            ("/2/users/by", "usernames=a,b", "users.read", "GET /2/users/by", "user.read", 2),
-            (
-                "/2/users/123/followers",
-                "max_results=1000",
-                "follows.read",
-                "GET /2/users/{id}/followers",
-                "following_followers.read",
-                1000,
-            ),
-            (
-                "/2/users/123/following",
-                "max_results=1000",
-                "follows.read",
-                "GET /2/users/{id}/following",
-                "following_followers.read",
-                1000,
-            ),
-            (
-                "/2/spaces",
-                "ids=1,2",
-                "space.read",
-                "GET /2/spaces",
-                "space.read",
-                2,
-            ),
-            (
-                "/2/tweets/analytics",
-                "ids=1,2",
-                "tweet.read",
-                "GET /2/tweets/analytics",
-                "analytics.read",
-                2,
-            ),
-            (
-                "/2/tweets/123/quote_tweets",
-                "max_results=100",
-                "tweet.read",
-                "GET /2/tweets/{id}/quote_tweets",
-                "posts.read",
-                100,
-            ),
-            (
-                "/2/tweets/123/retweeted_by",
-                "max_results=100",
-                "tweet.read",
-                "GET /2/tweets/{id}/retweeted_by",
-                "user.read",
-                100,
-            ),
-            (
-                "/2/users/123/tweets",
-                "max_results=100",
-                "users.read",
-                "GET /2/users/{id}/tweets",
-                "posts.read",
-                100,
-            ),
-            (
-                "/2/users/123/liked_tweets",
-                "max_results=5",
-                "like.read",
-                "GET /2/users/{id}/liked_tweets",
-                "posts.read",
-                5,
-            ),
-            (
-                "/2/lists/123/members",
-                "max_results=100",
-                "list.read",
-                "GET /2/lists/{id}/members",
-                "list.read",
-                100,
-            ),
-            (
-                "/2/dm_events",
-                "max_results=100",
-                "dm.read",
-                "GET /2/dm_events",
-                "dm_event.read",
-                100,
-            ),
-        ],
-    )
-    def test_x_json_parse_error_with_request_hints_uses_fallback_without_error_log(
-        self,
-        tmp_path,
-        real_flow,
-        path,
-        query,
-        permission,
-        rule,
-        expected_category,
-        expected_quantity,
-    ):
-        """Recoverable parser failures should bill from hints without lost-visibility logs."""
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            path=path,
-            query=query,
-            permission=permission,
-            rule=rule,
-        )
-        flow.metadata["x_json_state"] = {
-            "body_parsed": False,
-            "body_truncated": False,
-            "parse_error": "incomplete json",
-        }
-        proxy_log = tmp_path / "proxy.jsonl"
-
-        p = self._call_and_get_single_billing(flow)
-
-        assert p["category"] == expected_category
-        assert p["quantity"] == expected_quantity
-        assert proxy_log.exists()
-        entries = [json.loads(line) for line in proxy_log.read_text().splitlines()]
-        assert all(entry["level"] != "error" for entry in entries)
-        assert all("unparseable" not in entry["message"].lower() for entry in entries)
-        assert all("parse_error" not in entry for entry in entries)
-
-    def test_x_json_parse_error_with_zero_max_results_is_no_reliable_hint(
-        self, tmp_path, real_flow
-    ):
-        """A zero max_results hint should preserve lost-visibility logs."""
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            path="/2/tweets/search/recent",
-            query="query=hello&max_results=0",
-            rule="GET /2/tweets/search/recent",
-        )
-        flow.metadata["x_json_state"] = {
-            "body_parsed": False,
-            "body_truncated": False,
-            "parse_error": "incomplete json",
-        }
-        proxy_log = tmp_path / "proxy.jsonl"
-
-        assert self._call_and_get_billing(flow) == []
-        entry = self._assert_lost_visibility_error(proxy_log)
-        assert entry["parse_error"] == "incomplete json"
-
-    @pytest.mark.parametrize(
-        ("path", "query", "permission", "rule"),
-        [
-            (
-                "/2/tweets/search/recent",
-                "query=hello&max_results=-5",
-                "tweet.read",
-                "GET /2/tweets/search/recent",
-            ),
-            (
-                "/2/tweets/search/recent",
-                "query=hello&max_results=999999",
-                "tweet.read",
-                "GET /2/tweets/search/recent",
-            ),
-            (
-                "/2/tweets/search/recent",
-                "query=hello&max_results=+50",
-                "tweet.read",
-                "GET /2/tweets/search/recent",
-            ),
-            (
-                "/2/tweets/search/recent",
-                "query=hello&max_results=%2050",
-                "tweet.read",
-                "GET /2/tweets/search/recent",
-            ),
-            (
-                "/2/lists/123/members",
-                "max_results=101",
-                "list.read",
-                "GET /2/lists/{id}/members",
-            ),
-            (
-                "/2/users/123/tweets",
-                "max_results=4",
-                "users.read",
-                "GET /2/users/{id}/tweets",
-            ),
-            (
-                "/2/users/search",
-                "query=alice&max_results=1001",
-                "users.read",
-                "GET /2/users/search",
-            ),
-        ],
-    )
-    def test_x_json_parse_error_with_invalid_max_results_preserves_audit_log(
-        self, tmp_path, real_flow, path, query, permission, rule
-    ):
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            path=path,
-            query=query,
-            body=b"not json",
-            permission=permission,
-            rule=rule,
-        )
-        proxy_log = tmp_path / "proxy.jsonl"
-
-        assert self._call_and_get_billing(flow) == []
-        self._assert_lost_visibility_error(proxy_log)
-
-    def test_x_json_parse_error_ignores_max_results_on_irrelevant_path(self, tmp_path, real_flow):
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            path="/2/tweets/123",
-            query="max_results=1000",
-            body=b"not json",
-            rule="GET /2/tweets/{id}",
-        )
-        proxy_log = tmp_path / "proxy.jsonl"
-
-        assert self._call_and_get_billing(flow) == []
-        self._assert_lost_visibility_error(proxy_log)
-
-    @pytest.mark.parametrize(
-        ("path", "query", "permission", "rule"),
-        [
-            ("/2/users/by", "ids=1,2", "users.read", "GET /2/users/by"),
-            (
-                "/2/tweets/search/recent",
-                "query=hello&ids=1,2",
-                "tweet.read",
-                "GET /2/tweets/search/recent",
-            ),
-        ],
-    )
-    def test_x_json_parse_error_ignores_id_like_hints_on_irrelevant_paths(
-        self, tmp_path, real_flow, path, query, permission, rule
-    ):
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            path=path,
-            query=query,
-            body=b"not json",
-            permission=permission,
-            rule=rule,
-        )
-        proxy_log = tmp_path / "proxy.jsonl"
-
-        assert self._call_and_get_billing(flow) == []
-        self._assert_lost_visibility_error(proxy_log)
-
-    @pytest.mark.parametrize(
-        ("path", "query", "permission", "rule"),
-        [
-            (
-                "/2/tweets",
-                f"ids={','.join(str(i) for i in range(101))}",
-                "tweet.read",
-                "GET /2/tweets",
-            ),
-            (
-                "/2/users/by",
-                f"usernames={','.join(f'user{i}' for i in range(101))}",
-                "users.read",
-                "GET /2/users/by",
-            ),
-        ],
-    )
-    def test_x_json_parse_error_ignores_excessive_selector_counts(
-        self, tmp_path, real_flow, path, query, permission, rule
-    ):
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            path=path,
-            query=query,
-            body=b"not json",
-            permission=permission,
-            rule=rule,
-        )
-        proxy_log = tmp_path / "proxy.jsonl"
-
-        assert self._call_and_get_billing(flow) == []
-        self._assert_lost_visibility_error(proxy_log)
-
-    @pytest.mark.parametrize(
-        ("query", "expected_quantity"),
-        [
-            ("ids=1,,2", 2),
-            ("ids=,1,2,", 2),
-            ("ids=1,+,2", 2),
-        ],
-    )
-    def test_billable_counts_fallback_filters_empty_id_segments(
-        self, tmp_path, real_flow, query, expected_quantity
-    ):
-        """body unparseable but ?ids= present -> uses non-empty id segment count."""
-        flow = self._make_x_flow(real_flow, tmp_path, query=query, body=b"not json")
-        p = self._call_and_get_single_billing(flow)
-        assert p["category"] == "posts.read"
-        assert p["quantity"] == expected_quantity
-
-    def test_billable_counts_fallback_filters_empty_username_segments(self, tmp_path, real_flow):
-        """body unparseable but ?usernames= present -> uses non-empty username count."""
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            path="/2/users/by",
-            query="usernames=a,,b",
-            body=b"not json",
-            permission="users.read",
-            rule="GET /2/users/by",
-        )
-        p = self._call_and_get_single_billing(flow)
-        assert p["category"] == "user.read"
-        assert p["quantity"] == 2
-
-    def test_billable_counts_fallback_empty_id_segments_are_no_hint(self, tmp_path, real_flow):
-        """body unparseable and only empty ?ids= segments -> no billing, with audit log."""
-        flow = self._make_x_flow(real_flow, tmp_path, query="ids=,,", body=b"not json")
-        proxy_log = tmp_path / "proxy.jsonl"
-        assert self._call_and_get_billing(flow) == []
-        assert proxy_log.exists()
-        content = proxy_log.read_text()
-        assert "unparseable" in content.lower()
-        assert '"level":"error"' in content or '"level": "error"' in content
-
-    def test_billable_counts_fallback_uses_path_scoped_max_results(self, tmp_path, real_flow):
-        """body unparseable on a max_results endpoint -> uses max_results."""
-        flow = self._make_x_flow(
-            real_flow,
-            tmp_path,
-            path="/2/tweets/search/recent",
-            query="query=hello&max_results=50",
-            body=b"not json",
-            rule="GET /2/tweets/search/recent",
-        )
-        p = self._call_and_get_single_billing(flow)
-        assert p["category"] == "posts.read"
-        assert p["quantity"] == 50
-
-    def test_skips_on_server_error(self, tmp_path, real_flow):
-        flow = self._make_x_flow(real_flow, tmp_path, status=500)
-        assert self._call_and_get_billing(flow) == []
-
-    def test_skips_on_rate_limit(self, tmp_path, real_flow):
-        flow = self._make_x_flow(real_flow, tmp_path, status=429)
-        assert self._call_and_get_billing(flow) == []
-
-    def test_skips_on_empty_permission(self, tmp_path, real_flow):
-        """Unknown-endpoint-allow has no stable pricing key."""
-        flow = self._make_x_flow(real_flow, tmp_path, permission="")
-        assert self._call_and_get_billing(flow) == []
-
-    def test_skips_on_empty_run_id(self, tmp_path, real_flow):
-        flow = self._make_x_flow(real_flow, tmp_path)
-        assert self._call_and_get_billing(flow, run_id="") == []
-
-    def test_skips_when_not_billable(self, tmp_path, real_flow):
-        """Firewalls with firewall_billable=False are not reported."""
-        flow = self._make_x_flow(real_flow, tmp_path)
-        flow.metadata["firewall_billable"] = False
-        assert self._call_and_get_billing(flow) == []
-
-    def test_skips_when_no_response(self, tmp_path, real_flow):
-        flow = self._make_x_flow(real_flow, tmp_path)
-        flow.response = None
-        assert self._call_and_get_billing(flow) == []
-
-    def test_skips_webhook_without_sandbox_token(self, tmp_path, real_flow):
-        """When sandbox token is empty, no webhook is enqueued."""
-        body = json.dumps({"data": {"id": "1", "text": "hi"}}).encode()
-        flow = self._make_x_flow(
-            real_flow, tmp_path, path="/2/tweets/1", body=body, rule="GET /2/tweets/{id}"
-        )
-        flow.metadata["vm_sandbox_token"] = ""
-        assert self._call_and_get_billing(flow) == []
