+++ crates/runner/mitm-addon/tests/x_connector_usage/__init__.py
@@ -1,0 +1,1 @@
+"""Direct X connector usage reporting tests."""
